+++ streamsx/.toolkit/com.ibm.streamsx.topology/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changes
 ==========
 
+## v1.16.0
+* [#1679](https://github.com/IBMStreams/streamsx.topology/issues/1679) Nested tuple: Support tuple as attribute type
+* [#2551](https://github.com/IBMStreams/streamsx.topology/issues/2551) Adapt streamsx for REST API changes
+* [#2527](https://github.com/IBMStreams/streamsx.topology/issues/2527) Python: Support tumbling window based on punctuation
+* [#2525](https://github.com/IBMStreams/streamsx.topology/issues/2525) Python: Provide punct_count() in TopologyTester to test the number of received punctuations
+* [#2520](https://github.com/IBMStreams/streamsx.topology/issues/2520) Python: Ability to handle received punctuations in for_each and streamsx.spl.spl.PrimitiveOperator
+* [#2518](https://github.com/IBMStreams/streamsx.topology/issues/2518) Python: Ability to submit punctuation in streamsx.spl.spl.PrimitiveOperator
+* [#2517](https://github.com/IBMStreams/streamsx.topology/issues/2517) Python: Provide Stream.punctor()
+* [#2516](https://github.com/IBMStreams/streamsx.topology/issues/2516) Python: Write punctuation markers with Stream.print() to stdout
+
 ## v1.15.10
 * [#2561](https://github.com/IBMStreams/streamsx.topology/issues/2561) Third-party lib updated to resolve security vulnerabilities (commons-compress: 1.19)